Handle player states better
Added more events and add a ticker to watch for player flying, falling and so on. Idk a better way to watch / handle for this?

We added a delay on trying to summon as well.

We will add an option to let the user decide if they want the pet dismissed when stealth.